+++ packages/mask/src/plugins/Trader/SNSAdaptor/trader/TradeForm.tsx
@@ -2,45 +2,23 @@ import { memo, useCallback, useMemo, useRef, useState } from 'react'
 import { useUpdateEffect } from 'react-use'
 import { first, noop } from 'lodash-es'
 import { BigNumber } from 'bignumber.js'
-import { useI18N } from '../../../../utils/index.js'
-import {
-    PluginWalletStatusBar,
-    SelectTokenChip,
-    TokenSecurityBar,
-    useSelectAdvancedSettings,
-    useTokenSecurity,
-    WalletConnectedBoundary,
-    EthereumERC20TokenApprovedBoundary,
-    ChainBoundary,
-} from '@masknet/shared'
-import { makeStyles, MaskColorVar, ActionButton } from '@masknet/theme'
+import { SelectTokenChip, TokenSecurityBar, useSelectAdvancedSettings, useTokenSecurity } from '@masknet/shared'
+import { makeStyles, MaskColorVar } from '@masknet/theme'
 import { InputTokenPanel } from './InputTokenPanel.js'
 import { alpha, Box, chipClasses, Collapse, IconButton, lighten, Typography } from '@mui/material'
-import {
-    ChainId,
-    formatWeiToEther,
-    GasConfig,
-    GasEditor,
-    SchemaType,
-    Transaction,
-    ZERO_ADDRESS,
-} from '@masknet/web3-shared-evm'
-import { isLessThan, rightShift, multipliedBy, leftShift } from '@masknet/web3-shared-base'
+import { ChainId, GasConfig, GasEditor, Transaction } from '@masknet/web3-shared-evm'
+import { rightShift, multipliedBy, leftShift } from '@masknet/web3-shared-base'
 import { Tune as TuneIcon } from '@mui/icons-material'
 import { TokenPanelType, TradeInfo } from '../../types/index.js'
 import { Icons } from '@masknet/icons'
-import { isNativeTokenWrapper } from '../../helpers/index.js'
+import { useI18N } from '../../../../utils/index.js'
 import { DefaultTraderPlaceholder, TraderInfo } from './TraderInfo.js'
-import { MINIMUM_AMOUNT, MIN_GAS_LIMIT } from '../../constants/index.js'
-import { resolveTradeProviderName } from '../../pipes.js'
-import { useTradeApproveComputed } from '../../trader/useTradeApproveComputed.js'
-import { isDashboardPage, isPopupPage, PopupRoutes, PluginID, NetworkPluginID } from '@masknet/shared-base'
+import { MIN_GAS_LIMIT } from '../../constants/index.js'
+import { isDashboardPage, isPopupPage, PluginID, NetworkPluginID } from '@masknet/shared-base'
 import { useChainContext, useNetworkContext, useWeb3State } from '@masknet/web3-hooks-base'
 import { AllProviderTradeContext } from '../../trader/useAllProviderTradeContext.js'
-import { TokenSecurityBoundary } from '@masknet/plugin-go-plus-security'
 import { currentSlippageSettings } from '../../settings.js'
 import { PluginTraderMessages } from '../../messages.js'
-import Services from '../../../../extension/service.js'
 import { useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra/content-script'
 import { useIsMinimalModeDashBoard } from '@masknet/plugin-infra/dashboard'
 import type { Web3Helper } from '@masknet/web3-helpers'
@@ -117,12 +95,7 @@ const useStyles = makeStyles<{
         section: {
             width: '100%',
         },
-        button: {
-            borderRadius: 8,
-        },
-        disabledButton: {
-            borderRadius: 8,
-        },
+
         selectedTokenChip: {
             borderRadius: '22px!important',
             height: 'auto',
@@ -179,21 +152,7 @@ const useStyles = makeStyles<{
         whiteDrop: {
             color: '#ffffff !important',
         },
-        stateBar: {
-            position: 'sticky',
-            bottom: 0,
-            boxShadow: `0px 0px 20px ${alpha(
-                theme.palette.maskColor.shadowBottom,
-                theme.palette.mode === 'dark' ? 0.12 : 0.05,
-            )}`,
-        },
-        unlockContainer: {
-            margin: 0,
-            width: '100%',
-            ['& > div']: {
-                padding: '0px !important',
-            },
-        },
+
         title: {
             fontSize: 14,
             fontWeight: 700,
@@ -216,7 +175,6 @@ export interface AllTradeFormProps extends withClasses<'root'> {
     focusedTrade?: TradeInfo
     gasPrice?: string
     onFocusedTradeChange: (trade: TradeInfo) => void
-    onSwap: () => void
     onSwitch: () => void
     settings?: boolean
     gasConfig?: GasConfig
@@ -233,7 +191,7 @@ export const TradeForm = memo<AllTradeFormProps>(
         inputTokenBalance,
         focusedTrade,
         onFocusedTradeChange,
-        onSwap,
+
         gasPrice,
         onSwitch,
         settings,
@@ -249,28 +207,13 @@ export const TradeForm = memo<AllTradeFormProps>(
         const { chainId } = useChainContext()
         const { pluginID } = useNetworkContext()
         const { Others } = useWeb3State()
-        const { isSwapping, allTradeComputed } = AllProviderTradeContext.useContainer()
+        const { allTradeComputed } = AllProviderTradeContext.useContainer()
         const [isExpand, setExpand] = useState(false)
 
         const snsAdaptorMinimalPlugins = useActivatedPluginsSNSAdaptor(true)
         const isSNSClosed = snsAdaptorMinimalPlugins?.map((x) => x.ID).includes(PluginID.GoPlusSecurity)
         const isDashboardClosed = useIsMinimalModeDashBoard(PluginID.GoPlusSecurity)
         const isTokenSecurityEnable = !isSNSClosed && !isDashboardClosed
-
-        const { value: tokenSecurityInfo, error } = useTokenSecurity(
-            pluginID === NetworkPluginID.PLUGIN_EVM ? (chainId as ChainId) : undefined,
-            outputToken?.address.trim(),
-            isTokenSecurityEnable,
-        )
-
-        const isRisky = tokenSecurityInfo?.is_high_risk
-
-        // #region approve token
-        const { approveToken, approveAmount, approveAddress } = useTradeApproveComputed(
-            focusedTrade?.value ?? null,
-            focusedTrade?.provider,
-            inputToken,
-        )
 
         // #region token balance
         const inputTokenBalanceAmount = new BigNumber(inputTokenBalance || '0')
@@ -291,51 +234,6 @@ export const TradeForm = memo<AllTradeFormProps>(
             amount_ = Others?.isNativeTokenSchemaType(inputToken?.schema) ? amount_.minus(gasFee) : amount_
             return leftShift(BigNumber.max(0, amount_), inputToken?.decimals).toFixed(5)
         }, [focusedTrade, gasPrice, inputTokenTradeAmount, inputToken, Others?.isNativeTokenSchemaType])
-
-        // #region UI logic
-        // validate form return a message if an error exists
-        const validationMessage = useMemo(() => {
-            if (inputTokenTradeAmount.isZero()) return t('plugin_trader_error_amount_absence')
-            if (isLessThan(inputAmount, MINIMUM_AMOUNT)) return t('plugin_trade_error_input_amount_less_minimum_amount')
-            if (!inputToken || !outputToken) return t('plugin_trader_error_amount_absence')
-            if (!trades.length) return t('plugin_trader_error_insufficient_lp')
-
-            if (
-                inputTokenBalanceAmount.isLessThan(inputTokenTradeAmount) ||
-                (Others?.isNativeTokenSchemaType(inputToken.schema) &&
-                    formatWeiToEther(inputTokenTradeAmount).isGreaterThan(maxAmount))
-            )
-                return t('plugin_trader_error_insufficient_balance', {
-                    symbol: inputToken?.symbol,
-                })
-
-            if (focusedTrade?.value && !focusedTrade.value.outputAmount) return t('plugin_trader_no_enough_liquidity')
-            return ''
-        }, [
-            inputAmount,
-            focusedTrade,
-            trades,
-            inputToken,
-            outputToken,
-            inputTokenBalanceAmount.toFixed(),
-            inputTokenTradeAmount.toFixed(),
-            maxAmount,
-            Others?.isNativeTokenSchemaType,
-        ])
-        // #endregion
-
-        // #region native wrap message
-        const nativeWrapMessage = useMemo(() => {
-            if (focusedTrade?.value) {
-                if (isNativeTokenWrapper(focusedTrade.value)) {
-                    return focusedTrade.value.trade_?.isWrap ? t('plugin_trader_wrap') : t('plugin_trader_unwrap')
-                }
-                return t('plugin_trader_swap_amount_symbol')
-            } else {
-                return t('plugin_trader_no_trade')
-            }
-        }, [focusedTrade, outputToken])
-        // #endregion
 
         const handleAmountChange = useCallback(
             (amount: string) => {
@@ -432,14 +330,12 @@ export const TradeForm = memo<AllTradeFormProps>(
         }, [chainId, focusedTrade?.gas.value, selectAdvancedSettings, gasConfig])
         // #endregion
 
-        // #region if `isPopup` be true, click the plugin status bar need to  open popup window
-        const openSelectWalletPopup = useCallback(() => {
-            Services.Helper.openPopupWindow(PopupRoutes.SelectWallet, {
-                chainId,
-            })
-        }, [chainId])
+        const { value: tokenSecurityInfo, error } = useTokenSecurity(
+            pluginID === NetworkPluginID.PLUGIN_EVM ? (chainId as ChainId) : undefined,
+            outputToken?.address.trim(),
+            isTokenSecurityEnable,
+        )
 
-        // #endregion
         return (
             <>
                 <Box className={classes.root}>
@@ -549,136 +445,6 @@ export const TradeForm = memo<AllTradeFormProps>(
                         </Box>
                     </Box>
                 ) : null}
-                <Box className={classes.stateBar}>
-                    <PluginWalletStatusBar onClick={isPopup ? openSelectWalletPopup : undefined}>
-                        {/* TODO: remove the chain boundary after sol network dex be added */}
-                        {settings && pluginID !== NetworkPluginID.PLUGIN_EVM ? (
-                            <ChainBoundary
-                                expectedPluginID={NetworkPluginID.PLUGIN_EVM}
-                                expectedChainId={chainId as ChainId}>
-                                <WalletConnectedBoundary offChain>
-                                    <EthereumERC20TokenApprovedBoundary
-                                        onlyInfiniteUnlock
-                                        spender={approveAddress}
-                                        amount={approveAmount.toFixed()}
-                                        classes={{ container: classes.unlockContainer }}
-                                        contractName={
-                                            focusedTrade?.provider
-                                                ? resolveTradeProviderName(focusedTrade.provider)
-                                                : ''
-                                        }
-                                        infiniteUnlockContent={t('plugin_trader_unlock_symbol', {
-                                            symbol: approveToken?.symbol,
-                                        })}
-                                        expectedChainId={chainId as ChainId}
-                                        token={
-                                            !isNativeTokenWrapper(focusedTrade?.value ?? null) &&
-                                            approveToken?.schema === SchemaType.ERC20 &&
-                                            !!approveAmount.toNumber()
-                                                ? approveToken
-                                                : undefined
-                                        }
-                                        ActionButtonProps={{
-                                            color: 'primary',
-                                            style: { borderRadius: 8 },
-                                            size: 'medium',
-                                        }}>
-                                        <TokenSecurityBoundary
-                                            tokenInfo={{
-                                                name: tokenSecurityInfo?.token_name ?? '--',
-                                                chainId: tokenSecurityInfo?.chainId ?? ChainId.Mainnet,
-                                                contract: tokenSecurityInfo?.contract ?? ZERO_ADDRESS,
-                                            }}
-                                            disabled={
-                                                focusedTrade?.loading ||
-                                                !focusedTrade?.value ||
-                                                !!validationMessage ||
-                                                isSwapping
-                                            }
-                                            onSwap={onSwap}
-                                            showTokenSecurity={isTokenSecurityEnable && isRisky}>
-                                            <ActionButton
-                                                fullWidth
-                                                loading={isSwapping}
-                                                variant="contained"
-                                                disabled={
-                                                    focusedTrade?.loading ||
-                                                    !focusedTrade?.value ||
-                                                    !!validationMessage ||
-                                                    isSwapping
-                                                }
-                                                classes={{ root: classes.button, disabled: classes.disabledButton }}
-                                                color="primary"
-                                                onClick={onSwap}>
-                                                {validationMessage || nativeWrapMessage}
-                                            </ActionButton>
-                                        </TokenSecurityBoundary>
-                                    </EthereumERC20TokenApprovedBoundary>
-                                </WalletConnectedBoundary>
-                            </ChainBoundary>
-                        ) : (
-                            <WalletConnectedBoundary offChain>
-                                <EthereumERC20TokenApprovedBoundary
-                                    onlyInfiniteUnlock
-                                    spender={approveAddress}
-                                    amount={approveAmount.toFixed()}
-                                    classes={{ container: classes.unlockContainer }}
-                                    contractName={
-                                        focusedTrade?.provider ? resolveTradeProviderName(focusedTrade.provider) : ''
-                                    }
-                                    infiniteUnlockContent={t('plugin_trader_unlock_symbol', {
-                                        symbol: approveToken?.symbol,
-                                    })}
-                                    expectedChainId={
-                                        pluginID === NetworkPluginID.PLUGIN_EVM ? (chainId as ChainId) : undefined
-                                    }
-                                    token={
-                                        !isNativeTokenWrapper(focusedTrade?.value ?? null) &&
-                                        approveToken?.schema === SchemaType.ERC20 &&
-                                        !!approveAmount.toNumber()
-                                            ? approveToken
-                                            : undefined
-                                    }
-                                    ActionButtonProps={{
-                                        color: 'primary',
-                                        style: { borderRadius: 8 },
-                                        size: 'medium',
-                                    }}>
-                                    <TokenSecurityBoundary
-                                        tokenInfo={{
-                                            name: tokenSecurityInfo?.token_name ?? '--',
-                                            chainId: tokenSecurityInfo?.chainId ?? ChainId.Mainnet,
-                                            contract: tokenSecurityInfo?.contract ?? ZERO_ADDRESS,
-                                        }}
-                                        disabled={
-                                            focusedTrade?.loading ||
-                                            !focusedTrade?.value ||
-                                            !!validationMessage ||
-                                            isSwapping
-                                        }
-                                        onSwap={onSwap}
-                                        showTokenSecurity={isTokenSecurityEnable && isRisky}>
-                                        <ActionButton
-                                            fullWidth
-                                            loading={isSwapping}
-                                            variant="contained"
-                                            disabled={
-                                                focusedTrade?.loading ||
-                                                !focusedTrade?.value ||
-                                                !!validationMessage ||
-                                                isSwapping
-                                            }
-                                            classes={{ root: classes.button, disabled: classes.disabledButton }}
-                                            color="primary"
-                                            onClick={onSwap}>
-                                            {validationMessage || nativeWrapMessage}
-                                        </ActionButton>
-                                    </TokenSecurityBoundary>
-                                </EthereumERC20TokenApprovedBoundary>
-                            </WalletConnectedBoundary>
-                        )}
-                    </PluginWalletStatusBar>
-                </Box>
             </>
         )
     },
